% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{alphaplot_pp}
\alias{alphaplot_pp}
\title{Alpha Diversity Plot}
\usage{
alphaplot_pp(richtable, xvar, yvar, wrapvar = NULL, colno = NULL)
}
\arguments{
\item{colno}{}
}
\description{
Alpha Diversity Plot
}
