% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_percentage_assigned}
\alias{get_percentage_assigned}
\title{Calculate the percentage of assigned taxa at each taxonomic rank}
\usage{
get_percentage_assigned(ps)
}
\arguments{
\item{ps}{A phyloseq object.}
}
\value{
A data frame containing the percentage of assigned taxa at each taxonomic rank.
}
\description{
This function calculates the percentage of assigned taxa at each taxonomic rank
based on a given phyloseq object.
}
\examples{
library(phyloseq)
data(GlobalPatterns)
get_percentage_assigned(GlobalPatterns)

}
