% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{melt_to_top_n}
\alias{melt_to_top_n}
\title{Melt to top n taxa
Takes a phyloseq object and returns a melted table with only the top n taxa labelled.
Taxa with assignments at the specified rank that are not in the top n will be labelled "Other".
Those without assignments will be labelled "Unassigned"}
\usage{
melt_to_top_n(ps, n, rank)
}
\arguments{
\item{rank}{}
}
\description{
Melt to top n taxa
Takes a phyloseq object and returns a melted table with only the top n taxa labelled.
Taxa with assignments at the specified rank that are not in the top n will be labelled "Other".
Those without assignments will be labelled "Unassigned"
}
