% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{merge_orientations}
\alias{merge_orientations}
\title{Merge Two Differently oriented DADA2 runs from the same samples}
\usage{
merge_orientations(dada_fwd, dada_rvs)
}
\arguments{
\item{dada_rvs}{}
}
\description{
Merge Two Differently oriented DADA2 runs from the same samples
}
