% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_taxa_abundance}
\alias{plot_taxa_abundance}
\title{Plot Taxa Abundance}
\usage{
plot_taxa_abundance(
  ps,
  rank,
  xsep,
  wrap = NULL,
  n = 20,
  colno = NULL,
  byabundance = TRUE
)
}
\arguments{
\item{byabundance}{}
}
\description{
Plot Taxa Abundance
}
