% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ps_from_ampliseq}
\alias{ps_from_ampliseq}
\title{DADA2 output to physloseq object}
\usage{
ps_from_ampliseq(directory, metadata = NULL, ranks)
}
\arguments{
\item{ranks}{}
}
\description{
DADA2 output to physloseq object
}
